feat: move button component style to theme package
