+++ packages/components/button/package.json
@@ -35,6 +35,7 @@
   },
   "dependencies": {
     "@consolelabs/icons": "workspace:*",
+    "@consolelabs/theme": "workspace:*",
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.0.0"
   },
