+++ packages/components/button/tsconfig.json
@@ -1,5 +1,5 @@
 {
   "extends": "tsconfig/react-library.json",
-  "include": ["src", "index.ts"],
+  "include": ["src", "stories", "index.ts"],
   "exclude": ["dist", "node_modules"]
 }
