+++ packages/theme/src/components/index.ts
@@ -2,6 +2,7 @@ export * from './accordion'
 export * from './alert'
 export * from './avatar'
 export * from './badge'
+export * from './button'
 export * from './dropdown'
 export * from './input'
 export * from './input-field'
